+++ package/backend/sendEmail.server.ts
@@ -1,13 +1,14 @@
 import { SES } from "@aws-sdk/client-ses";
 import type React from "react";
 import ReactDOMServer from "react-dom/server";
+import mockEmailLocally from "./mockEmailLocally.server";
 
 const ses = new SES({
   endpoint: process.env.AWS_ENDPOINT,
 });
 export const supportEmail = "support@samepage.network";
 
-const sendEmail = ({
+const sendEmail = async ({
   to = supportEmail,
   body,
   subject,
@@ -22,18 +23,23 @@ const sendEmail = ({
 }): Promise<string> => {
   const Data =
     typeof body === "string" ? body : ReactDOMServer.renderToStaticMarkup(body);
+  const Body = {
+    Html: {
+      Charset: "UTF-8",
+      Data,
+    },
+  };
+  // if (process.env.NODE_ENV === "development") {
+  //   const MessageId = mockEmailLocally(Body);
+  //   return MessageId;
+  // }
   return ses
     .sendEmail({
       Destination: {
         ToAddresses: typeof to === "string" ? [to] : to,
       },
       Message: {
-        Body: {
-          Html: {
-            Charset: "UTF-8",
-            Data,
-          },
-        },
+        Body,
         Subject: {
           Charset: "UTF-8",
           Data: subject,
